Add Dataset::PlaceholderLiteralizer optimization framework
Dataset::PlaceholderLiteralizer is a new class that allows one
to record the changes made to a given dataset with placeholder
arguments.  The resulting dataset is then literalized, with the
placeholder arguments recording the current offset in the SQL
string insteading of literalizing themselves into the string.

Using the offsets, the string is split into an array of fragments.
Using that array, you can then pass arbitrary arguments into the
loader, and it can build the SQL string just by simple string
concatenation.

This is about 3 times faster in generating the SQL, and since
generating the SQL is a good percentage of the time spent for
simple queries that only return one row (assuming a local database
connection), this can speed up such queries by about 50%.

There are corner cases that this framework can't handle, but it
handles the vast majority of cases that Sequel users are likely to
use.  This includes things such as knowing that strings passed
to where should be considered literal strings:

  def find(arg)
    where(arg)
  end
  find("s = 1")  # WHERE (s = 1)
  find(:s=>1)    # WHERE (s = 1)

as well as where you are using the argument on the right hand side
of a conditions specifier, where the operator used changes based
on the value of the placeholder:

  def find_by_id(arg)
    where(:id=>arg)
  end
  find_by_id(1)      # WHERE (id = 1)
  find_by_id([1, 2]) # WHERE (id IN (1, 2))

This framework is similar to but more flexible than prepared
statements, since the underlying SQL can change from call to call.