Add Model.finder for creating optimized finder methods using Dataset::PlaceholderLiteralizer
This makes it easier to use Dataset::PlaceholderLiteralizer for
models, providing a simpler API.  You just pass it a symbol for
a model class method that returns a dataset, and it will create
an optimized method that will do the given type of action (default
type is :first, for returning the first record).

Add a Model.first_where optimized method automatically at runtime,
and have Model.find, .first, and .first! call it if given a single
argument.  This can speed up these methods by about 50% for local
database connections.

To avoid creating a bunch of unnecessary objects, the
PlaceholderLiteralizer objects are created lazily on an as needed
basis.