Make emulated operators on some adapters use an append-only approach
This adds Dataset#complex_expression_emulate_append private method
to the core, which contains implementations of the emulated code for
the %, >>, <<, &, |, ^, and B~ operators that were previously
duplicated in 7 of Sequel's adapters.

This changes Dataset#complex_expression_arg_pairs private method
to return nested expression objects instead of already literalized
objects, which may break backwards compatibility for external
adapters that call the method.