Add Dataset#disable_insert_returning on PostgreSQL for skipping implicit use of RETURNING
This method was removed in 3.35.  However, there are certain cases
even in PostgreSQL 9.3 where INSERT RETURNING doesn't work.  This
at least makes it easy to cope with such situations, though if
you use disable_insert_returning, you are responsible for manually
handling the primary key values of the inserted rows.  I'm not
going to bring back all of the code that guesses the sequence and
tries to use currval after the insert.