Add Dataset#current_datetime for returning current Time/DateTime object
The 3 plugins that needed a current datetime value were inconsistent,
with one caling Sequel.datetime_class.now and the other two calling
Time.now.  This makes sure that Sequel.datetime_class is used in
both places.  It also makes it possible to override the method to
change the behavior for all three of these plugins.