Added (trivial) support for the openedge.jar JDBC driver
For now, we don't extend the DB object or anything. We only return the driver name, so that we can run when the classloading is a bit messed up (which can be the case e.g. when running from `jruby-complete-*.jar`).

In the future, we could extend this to support the different database methods etc. if needed. For now, just being able to run straight SQL is "good enough" for me.