Further increase PostgreSQL array parsing performance (another 2x increase)
Thanks to strelec for the pointer to use the each_char method.
In addition to being over twice as fast, this also greatly
reduces memory consumption.

This loses the ability to detect some malformed arrays (those that
exit before the string is processed), but the performance
decrease is not worth trying to detect those.  There are plenty
of other malformed arrays where exceptions weren't raised
in either case.