Use INSERT VALUES with multiple rows for Dataset#import and #multi_insert on more databases
This adds support for using INSERT VALUES with multiple
rows to H2, HSQLDB, Derby, SQLAnywhere, CUBRID,
SQLite >=3.7.11, and MSSQL >=2008.

According to SQL-92, INSERT VALUES accepts multiple rows.
I really wanted to make using INSERT VALUES with multiple
rows the default, but unfortunately, support for it is not as
widespread as I'd like. It is missing completely in Oracle,
DB2, Firebird, and Access, and not available on MSSQL < 2008
or SQLite < 3.7.11.

This actually adds two new strategies to multi_insert_sql.
The first is VALUES with multiple rows.  The second is a UNION
based approach used on old versions of MSSQL and SQLite, which
should work on any database that supports INSERT with a SELECT
query, as long as SELECT doesn't require a FROM table (which
rules out DB2, Oracle, and Firebird).  The UNION approach can
be extended to handle those databases, but I'll wait for a user
to request that before implementing it.