Handle placeholder literal values when emulating offsets with row_number
When emulating offsets with row_number, we need to get the columns
for the dataset in order to construct the SQL for the dataset. This
is a little bit of a chicken and egg issue, since getting the columns
requires constructing SQL related to the dataset.  With the previous
code, the code to generate the columns would fail, since it wouldn't
have a value for the placeholder.  This works around that by using
NULL for the value.