Support :correlated_subquery as an eager_graph and filter by associations limit strategy for one_to_* associations
Sequel used to have a correlated_subquery approach for regular
eager loading, but that was removed in 4.0.  This uses a similar
approach for eager_graph and filter by associations.

In my testing I have found that the correlated subquery strategy
is fastest in some cases for eager_graph and filter by
associations.  This is usually when you are selecting a small
number of current records out of a large table of current
records, and the number of associated records for each current
record isn't too large.

I have no plans to add support for using correlated subqueries for
regular eager loading, as the current union approach is always
faster in my testing.

As the only previous filter by associations limit strategies were
distinct_on and window_function, filtering by limited associations
didn't work on databases that don't support window functions. This
commit allows them to work on such databases.

There are a couple cases where the correlated subquery strategy
won't work correctly. One is when you are using MySQL, which doesn't
allow LIMIT clauses inside subqueries passed to IN.  The other is
when you are using composite primary keys and the database doesn't
support multiple columns in IN, since in that case Sequel has to
emulate the IN using a separate query, and that doesn't work
for correlated subqueries.

Currently, the correlated_subquery strategy is only supported for
one_to_one and one_to_many, and probably only works correctly if
you are not using qualified identifiers when filtering or ordering
the association.  It is possible to handle other association types
and automatically requalify qualified identifiers, but I don't
plan on implementing that unless I get a request for it, as those
are significantly more complex.