Add Dataset#supports_limits_in_subqueries? and #offset_requires_count?
This allows us not to test for specific database types in the code
and specs.  The supports_limits_in_subqueries? replaces a specific
check for MySQL.  The offset_requires_count? checks are new, and
avoid using a UNION or correlated subquery approach automatically
on Access when the association uses an offset.