Replace Dataset#offset_requires_count? with #supports_offsets_in_correlated_subqueries?
It looks like offsets in correlated subqueries don't work in any
case where the offsets are emulated.  This is true both when
a count is used (Access), and also when row_number is used to
emulate, as the query to get the columns to use fails in that
case due to the correlation not being present.

This means that the correlated_subquery limit strategy is not going
to work in cases where an offset is used and the database emulates
offsets, so manually skip those specs.