Add Database#transaction :auto_savepoint option, for automatically using a savepoint in nested transactions
This is most useful when testing, where the test runner can use
the option, which ensures that rollbacks for internal transactions
still work correctly.

Unfortunately, this required some refactoring of the transaction
code, and it breaks backwards compatibility with any custom
adapters that use the :savepoint_level or :savepoints transaction
hash entry.  Instead of doing _trans(conn)[:savepoint_level], users
should switch to savepoint_level(conn).  The savepoint level
is set/modified in add_transaction and not begin_transaction now,
which should be easier for custom adapters.