Use placeholder literalizer when eager loading for up to 20% speedup
This only adds a significant speedup if you are not eager loading
many rows.  If you are eager loading many rows, the time spent
building the query is small compared to loading the associated
rows.

Currently, this caching works because eager_limit_strategy cannot
vary per-call.  If we move to allowing eager_limit_strategy to
change per-call, then we'll need to change the caching.