Use placeholder literalizer for regular association loading for up to 85% speedup
This uses a placeholder literalizer for regular association loading
if the association does not depend on values other than the keys
and a dynamic callback is not used.

The pg_array_associations plugin does not like this because it
doesn't use simple key values, so add a use_placheolder_loader
method as a deoptimization flag to turn this off.

The sharding plugin doesn't like this because the placeholder
loader doesn't respect the shard, so also offer the ability for
the model objects to deoptimize by having the
_associated_object_loader method return nil.

An explicit SQL::BooleanExpression.new is used here as the key
values should always be equal and it's significantly faster.