Work around MySQL 5.6+ bug when combining DROP FOREIGN KEY and DROP INDEX in same ALTER TABLE statement
I'm not sure exactly when this bug started, but at least on 5.6.17,
issuing a DROP FOREIGN KEY and a DROP INDEX in the same ALTER TABLE
statement, when DROP INDEX is dropping the unique index that the
foreign key depends on, now raises an error.  I'm assuming this
started in 5.6, so on 5.6+, always issue the DROP INDEX in a
separate ALTER TABLE statement.