Speed up Sequel.split_symbol 10-20x by caching results
Chances are, if a symbol needs to be split once, it needs to
be split more than once, so caching previous results should
significantly improve performance.

There is no cache expiration for this cache.  I don't think there
is a denial of service risk, as creating arbitrary symbols is
already a denial of service risk on ruby < 2.2, and even on ruby
2.2+, creating symbols from user input is a bad idea from a
security perspective.

From my testing, this speeds up dataset literalization by about
80% for some common datasets.  Unlike the previous commit, this
speedup does not drop as the dataset gets more complex, though
it does vary depending on the ratio of symbols the dataset has
to literalize versus other objects the dataset has to literalize.