Speed up Dataset#literal for symbols 60% by caching results, speeding up dataset literalization up to 40% or more
While splitting symbols was the more expensive part of
literalization, mangling and quoting isn't cheap either,
so caching the resulting literal output is a big
performance win.

Unlike the split_symbols case, how symbols are literalized depends
on both per-Database and per-Dataset settings.  Cache the literal
symbols in the Database object, don't use the cache if the
Dataset has overridden the Database's default settings or a
prepared statement is being used, and clear the cache if the
default settings change.