Refactor type conversion in the jdbc adapter, for up to a 20% speedup
This rips out the old type conversion code and replaces it with
a new implementation that is more similar to other adapters.

The previous code didn't know how to determine the conversion
proc to use before iterating over the results.  It would look
at the object returned for each column and guess at the
conversion proc to use based on the object's class.  It tried
to do this in such a way that the conversion proc was cached
on first lookup, but it was still very clumsy.

I think the reason it was setup this way is that I didn't know
the JDBC metadata API well enough.  Once I found out about how
java.sql.Types and ResultSetMetaData#getColumnType work, it was
fairly easy to make the jdbc adapter work like the other adapters
and set up the conversion procs before iterating.

In addition to setting up the conversion procs first, this
changes the jdbc adapter so that instead of calling getObject
for everything, we call more specific methods (getLong, getString,
etc.).  We can do this because we know the type in advance.

This keeps the convert_types setting for the jdbc adapter.  There
are two conversion maps for jdbc adapter Database objects, one
that tries the hardest to convert to the correct ruby type (the
default), and one that just does a simple conversion to a java
type (used when convert_types is false).

This makes many of the JDBC adapters much simpler.  Instead of
trying to recognize JDBC driver specific classes, they can generally
just use the standard API.  It's also fairly easy to extend the
standard API to handle database specific types or quirks of types.

This is probably not completely backwards compatible.  It passes the
specs, but I'm sure there are corner cases that are handled
differently.  Hopefully the new handling is better, but if this
causes regressions, I'll try to patch them quickly if I hear about
them, assuming it is possible to do so.

One known difference is that when dealing with a PostgreSQL array
that doesn't have a defined conversion proc, the entries in the array
will not be converted.  For example, if you retrieve a date[] type,
you'll end up with an array of java.sql.Date objects instead of ruby
Date objects.  The solution is to use the pg_array extension, which
adds a conversion proc for date[] instances.

While this is mostly being done because it is much simpler, it also
provides a modest performance improvement.

One small new feature here is that timestamp with local time zone
types are now returned correctly as Time/DateTime objects in the
jdbc/oracle adapter.