Automatically determine array type in pg_array_associations plugin, explicitly cast arrays in more places
This changes the plugin so that the array type is determined
automatically.  This should handle most cases, though you can still
use the :array_type association option to override the type.

This also extends the explicit casting of arrays to any cases where
arrays are used with literal values (where PostgreSQL may not be
able to determine the correct type).