Support PostgreSQL 9.4+ jsonb type in the pg_json extension
There are not implicit casts between the json and jsonb types,
so Sequel needs to basically separate the two and use different
classes for each.  So you have to use Sequel.pg_jsonb to get
a jsonb type.

Note that this does not include support for the new jsonb
functions in pg_json_ops.  For some reason, instead of keeping
the function names the same and just varying the type, they
had duplicate functions.  Hopefully this will change before the
official release of PostgreSQL 9.4.