Test jsonb bound variables in the postgres adapter specs
These tests originally only tested json, and wouldn't work
directly with jsonb.  However, the only thing they really need
to test is whether they can successfully used in bound variables,
so reduce the scope of the test by eliminating the queries
which depend on json<->text conversion, which leaves a spec
that works for both json and jsonb.