Add support for the PostgreSQL 9.4+ jsonb type to the pg_json_ops extension
This takes the support that previously only existed for the json
type, and extends it to the jsonb type, similar to the changes
made to the pg_json extension.

This does not add support for the new json/jsonb functions or
the new jsonb operators, that will come in a future commit.