Add :reset_implicit_qualifier option to Dataset#join_table, to set false to not reset the implicit qualifier
This makes it easier to do multiple joins in a star topology.
Sequel's default has always made it easier to do multiple joins
in a linear topology.  This option is easier than specifying the
:implicit_qualifier option to the first source table multiple
times.

This option name is too long, but I can't currently think of
a better one.