Use correct schema in schema parsing
It needs to default to "CURRENT_SCHEMA", not ""CURRENT_SCHEMA"" or
"'CURRENT_SCHEMA'" or "`CURRENT_SCHEMA`". This is why one of the
tests kept complaining about a 'tags.taggings_count' column which
clearly didn't exist; it existed on a tags table in another schema.
Good thing I've been using this db for other tests.

600 -> 600 examples
254 -> 238 failures