Minor modifications to nested_attributes reciprocal commit
Use add_reciprocal_object to set up the reciprocal, which means
that it will be setup only for reciprocal singular associations,
since other associations won't have a cached association array.

It's possible in the future that this can be extended to support
reciprocal plural associations, but I'm not sure what the corner
cases are in that case.