Use explicit columns when using automatically prepared SELECT statements in the prepared statement plugins
Previously, Sequel would use the same SQL in a prepared statement
as it would in a regular query, except for the placeholders.  In
most cases, this resulted in Sequel using SELECT * for SELECT
queries.  Unfortunately, databases such as PostgreSQL complain
for these types of queries if you are adding columns to a table.

This can cause problems when running migrations for a concurrently
running application that uses the prepared statements plugin. The
only previous alternatives were:

1) Don't use prepared statements while the migration code is running.
2) Shut down the application while the migration code is running.
3) Accept that some queries may fail while the migration code is
   running.

None of these alternatives were desirable.  With this change, you
can safely use the prepared statements plugin while running
migrations that add columns to existing tables.

Note that most other forms of schema modification will still cause
problems with a concurrently running application (such a renaming
a column), but those issues are not specific to the use of
prepared statements.