Add Model#set_nested_attributes to nested_attributes, allowing setting nested attributes options per-call
Previously, the default options set in the class-level
nested_attributes call were the only options that could be used.
This is obviously not friendly if you want to allow different
fields for nested attributes for different actions.

Note that this changes the internal nested_attributes API so
that instead of the reflection, a hash of metadata is passed
around.