Fix regression in timestamps and table inheritance plugin where column values would not be saved if validation is skipped (Fixes #839)
This adds a _before_validation private hook that is called both
before valid? and in save before validation is done, even if
validation is skipped.  This ensures the hook is always run
during save, and always run before validation.  It is useful
whenever plugin code needs to set column values on the object
that it might be validating, but must be set before save even
if not validating.

This regression occurred in 4.11.0, and since that isn't the
latest release, this will not be backported and a new gem
will not be released, per Sequel's usual policy.