Use all_tables and all_views for Database#tables and #views on Oracle
This works better for users with limited rights.  Similar to other
databases, we try to hide the system tables.  In the case of Oracle,
that's not all that simple to do, as there are a lot of system users,
so we may want to expand the IGNORE_OWNERS constant in the future.