Refactor fractional timestamp literalization
Add Dataset#timestamp_precision private method holding the number
of decimal digits of precision supported by timestamps in the
database.  Use this to automatically handle the formatting of
fractional seconds, avoiding the need to override
format_timestamp_usec in the mssql and sqlanywhere adapters.

While here, cleanup the specs so they don't depend on the current
date/time values, which simplifies them significantly.