Use associated model's existing selection for join associations if it consists solely of explicitly quailfied columns
For associations that don't require joins, the associated class's
dataset selection is used by default if the :select option is not
specified.  In most cases that is SELECT * since that is Sequel's
default.

That is not true for many_to_many and other associations
that don't require joins. Those have always used associated_table.*,
so that they do not automatically pick up columns in the join
table and so that they don't have any issues with columns becoming
ambiguous.

However, if the associated class's dataset consists of all explicitly
qualified columns, then it will not automatically pick up columns
in the join table, and there won't be any issues with columns
becoming ambiguous.  In that case, there is no reason to ignore the
model's default selection.