Add dataset_source_alias extension, for automatically aliasing datasets to their first source
I think this creates more usable aliases.  If not for backwards
compatibility, this would probably be the default behavior.

This should also help workaround issues when the column_aliases
plugin is used if you also use the model as an argument to
join_table, since then the model will use the dataset instead of
the table name when joining, which can change the alias used.