Qualify primary keys for models with joined datasets when looking up model instances by primary key
If the model uses a joined dataset, but also has a primary key,
assume the primary key is contained in the table specified by
the model's table_name, and use a qualified version of that
when looking up model instances.

This fixes issues when the primary key would be ambigious as
it is present in multiple tables in the model's dataset.