Refactor prepared insert_select support
This adds Dataset#insert_select_sql whereever #insert_select is
supported, and changes the prepared statement code to use that
instead of assuming returning is supported if insert_select is
supported.

This also adds integration tests for using insert_select with
bound variables and prepared statements.  Previously it was
tested indirectly via the prepared statement plugin support,
but it is better to also test support directly.

Additionally, this switches the adapters that implement
insert_select to all use with_sql_first to retrieve the row,
instead of either calling insert or manually trying to set
the SQL, which should be faster.

Remove the previous hack to the MSSQL adapter to get insert_select
to work.