Emulate RETURNING on Microsoft SQL Server using OUTPUT, as long as only simple column references are used
If only simple column references are used, you can take the
RETURNING values and automatically qualify them with INSERTED
or DELETED depending on the query, and they pretty much work
directly with the OUTPUT clause.

If both RETURNING and OUTPUT are specified, give OUTPUT priority.
If the returning clause is already present, have insert_select
use it instead of creating a new OUTPUT clause.