Make Model#save use insert_select if the dataset used for inserting already uses returning
Previously, if the model's dataset selected specific columns,
insert_select was not used.  This was because insert_select
used to automatically use RETURNING *, and if the dataset
did not SELECT *, using insert_select could change the columns
returned (e.g. turning a lazy loaded column into an eager loaded
column).

However, since insert_select now respects an existing
RETURNING clause, you can set the RETURNING clause appropriately,
so that insert_select wouldn't change the behavior, and then
it is beneficial to use as it reduces the number of queries.