Make bin/sequel add lib directory to front of $: instead of end
This fixes an issue where some code you run inside of bin/sequel
requires a gem that depends on sequel, in which case the gem
directory gets added before the lib directory.  Since the point
of this change to bin/sequel was to make make:

  bin/sequel

act like:

  ruby -I lib bin/sequel

The lib directory should go at the front of $:, since that is
what ruby -I does.