Make bin/sequel -v not exit if database/code string given
Previously, bin/sequel -v exited in all cases.  Now, it operates
more similarly to ruby -v, printing the version but continuing
if a database or code string is provided

I think it would be a good idea to extend this behavior so that
if code is piped in, it also executes it after printing the
version, but that is not currently ruby's behavior, so it shouldn't
be bin/sequel's either.