Unalias aliased tables in Dataset#insert
Some databases support aliased tables in UPDATE and DELETE, which
makes sense as they have WHERE clauses.  None that I could find
support aliasing the table in INSERT, so it should be safe to
unalias the table there.

This allows you to have a dataset with an aliased table and
insert, update, delete, and select using it, assuming your
database supports tables aliases in update and delete.