Fix regression in class_table_inheritance plugin when lazily loading column in middle table
This regression happened when you have a hierarchy such as:

 class A < Sequel::Model; plugin :class_table_inheritance; end
 class B < A; end
 class C < B; end

And you retrieve a partially loaded instance of the subclass
via the parent class:

  c = C.create
  c = A[c.id]

If you try to call a column method for a column contained in B's
table:

  c.column_in_table_b

You would get an error because it qualified the column with C's
table instead of B's table.

To fix this, options need to be added to lazy attributes getter
method creator so that you can define the dataset and qualifier
to use inside the getter.

class_table_inheritance can then use the options to setup the
lazy attributes getter properly.