Add documentation note about using symbols for column names on MySQL for certain schema changes
MySQL doesn't have the ability to change the column type, default,
or NULL/NOT NULL setting indepdently, it just has a CHANGE COLUMN
syntax that combines all three settings.  Sequel attempts to parse
the existing values from the schema, but that isn't going to work
unless the column is given as a symbol, since the schema hash is
keyed with a symbol.